Rename api-test packages to com.amido.stacks.tests
