+++ api-tests/src/test/java/com/amido/stacks/tests/api/WebServiceEndPoints.java
@@ -1,4 +1,4 @@
-package com.amido.stacks.api;
+package com.amido.stacks.tests.api;
 
 public enum WebServiceEndPoints {
   BASE_URL(System.getenv("BASE_URL")),
