+++ api-tests/src/test/java/com/amido/stacks/tests/api/pact/pacts/JavaAutomatedTestsMenuConsumer-JavaMenuAPI.json
