+++ api-tests/src/test/java/com/amido/stacks/tests/api/status/AppStatus.java
@@ -1,4 +1,4 @@
-package com.amido.stacks.api.status;
+package com.amido.stacks.tests.api.status;
 
 public enum AppStatus {
   RUNNING,
