+++ api-tests/src/test/java/com/amido/stacks/tests/api/templates/MergeFrom.java
@@ -1,4 +1,4 @@
-package com.amido.stacks.api.templates;
+package com.amido.stacks.tests.api.templates;
 
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
